Signout Api & Admin Dashboard

+++ actions/auth.actions.js
@@ -60,9 +60,20 @@ export const isUserLoggendIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-           type: authConstants.LOGOUT_REQUEST 
-        });
+        
+        dispatch({ type: authConstants.LOGOUT_REQUEST})
+        const res = await axios.post(`/admin/signout`)
+
+        if(res.status === 200){
+            localStorage.clear();
+            dispatch({type: authConstants.LOGOUT_SUCCESS});
+        }else{
+            dispatch({
+                type: authConstants.LOGIN_FAILURE,
+                payload: {error: res.data.error}
+            })
+        }
+
+        
     }
 }
